+++ crates/icp/src/agent.rs
@@ -22,10 +22,15 @@ pub struct Creator;
 #[async_trait]
 impl Create for Creator {
     async fn create(&self, id: Arc<dyn Identity>, url: &str) -> Result<Agent, CreateAgentError> {
-        let b = Agent::builder()
-            .with_url(url)
-            .with_arc_identity(id)
-            .with_ingress_expiry(Duration::from_secs(4 * MINUTE));
+        let mut b = Agent::builder().with_url(url).with_arc_identity(id);
+        let default_ingress_expiry = Duration::from_secs(4 * MINUTE);
+        if let Ok(ms) = std::env::var("ICP_CLI_TEST_ADVANCE_TIME_MS") {
+            b = b.with_ingress_expiry(
+                default_ingress_expiry + Duration::from_millis(ms.parse::<u64>().unwrap()),
+            );
+        } else {
+            b = b.with_ingress_expiry(default_ingress_expiry);
+        }
 
         Ok(b.build().context(AgentSnafu)?)
     }